solucion delete

+++ src/modules/deleteSeguro/deleteSeguro.service.js
@@ -3,20 +3,29 @@ const prisma = new PrismaClient();
 
 async function deleteSeguro(id) {
   try {
-    const result = await prisma.$transaction([
-      prisma.tiposeguro.deleteMany({
-        where: { segurocarro_id: parseInt(id) },
-      }),
-      prisma.seguroCarro.delete({
-        where: { id: parseInt(id) },
-      }),
-    ]);
+    const seguro = await prisma.SeguroCarro.findUnique({
+      where: { id: parseInt(id) },
+    });
+
+    if (!seguro) {
+      throw new Error('El seguro con el ID proporcionado no existe.');
+    }
+
+    await prisma.tiposeguro.deleteMany({
+      where: { segurocarro_id: parseInt(id) },
+    });
+    
+    const result = await prisma.SeguroCarro.delete({
+      where: { id: parseInt(id) },
+    });
+
     return result;
   } catch (error) {
+    console.error('Error al eliminar el seguro:', error);
     throw error;
   }
 }
 
 module.exports = {
   deleteSeguro,
-}; 
+};
